+++ apps/mobile/modules/gallery/index.ts
@@ -1,0 +1,4 @@
+/** biome-ignore-all lint/performance/noBarrelFile: go away */
+
+export { Gallery } from './src/module'
+export { GalleryOpenProps } from './src/types'